Skip snapshot testing of debug_log of delete_products e2e test
The debug_log of this specific test is too unpredictable, since it has code that deletes the products on purpose to make the test fail, the debug log entries can change due to a variety of factors, such as retries in E2E tests, etc. We still log the test failures, but we skip the PHP debug log.