Enhances PreCommand test case
Adds mocking capabilities for WordPress.org plugins and themes,
WooCommerce dependencies and download URLs, and generic URLs.

Introduces a mechanism to track and reset mocked data in the
application container, ensuring clean test environments.

Also introduces a mechanism to get access to the non normalized
environment information for more accurate tests.

Adds a test case for specifying a specific version of a WordPress.org
plugin as the SUT.